refactor: participant 패키지로 이동
