+++ src/main/java/blackjack/participant/Name.java
@@ -1,4 +1,4 @@
-package blackjack;
+package blackjack.participant;
 
 
 import utils.StringUtils;
