+++ src/main/java/blackjack/participant/Participant.java
@@ -1,4 +1,4 @@
-package blackjack;
+package blackjack.participant;
 
 import blackjack.card.Card;
